Move auto_webp feature into separate plugin
